Bugfix (reported by Anuj Kalia).
In a previous commit support for control rings has been added in the
shared memory segment, but the relevant function that calculated the shm
size was not updated accordingly.

This bug would not trigger a segfault when huge pages were used due to
enough headroom, but would fail with regular pages (4K). Fixed.