Fix issue 34: Add debug mode flag in CMakeLists.txt and also inline function.
The code changes add a debug mode flag in the CMakeLists.txt file to control whether the code is built in debug mode or not. The flag is set to 0 when debug mode is off. This change improves the flexibility of the build process and allows for easier debugging when needed.